arr function fix

+++ 0x0B-malloc_free/0-create_array.c
@@ -15,7 +15,7 @@ char *create_array(unsigned int size, char c)
 
 	if (size <= 0 || arr == NULL)
 		return (NULL);
-	while (x < size)
+	while (x < (int)size)
 	{
 		arr[x] = c;
 		x++;
